@@ -1,5 +1,5 @@
 {
-  "name": "FoodHub",
+  "name": "foodhub-front",
   "version": "0.0.1",
   "author": "Ionic Framework",
   "homepage": "https://ionicframework.com/",
@@ -13,54 +13,70 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^0.803.24",
+    "@angular-devkit/build-angular": "^0.803.28",
     "@angular/common": "~8.1.2",
-    "@angular/core": "~8.1.2",
+    "@angular/core": "^8.2.14",
     "@angular/forms": "~8.1.2",
+    "@angular/http": "^7.2.16",
     "@angular/platform-browser": "~8.1.2",
     "@angular/platform-browser-dynamic": "~8.1.2",
     "@angular/router": "~8.1.2",
     "@capacitor/core": "1.5.0",
-    "@ionic-native/core": "^5.21.1",
-    "@ionic-native/splash-screen": "^5.21.1",
-    "@ionic-native/status-bar": "^5.21.1",
-    "@ionic/angular": "^5.0.3",
+    "@ionic-native/core": "^5.27.0",
+    "@ionic-native/splash-screen": "^5.27.0",
+    "@ionic-native/status-bar": "^5.27.0",
+    "@ionic-native/stripe": "^5.27.0",
+    "@ionic/angular": "^5.3.1",
+    "@ionic/app-scripts": "^3.2.4",
     "@ionic/storage": "^2.2.0",
     "animate.css": "^3.7.2",
+    "capacitor": "^0.5.5",
     "chart.js": "^2.9.3",
-    "core-js": "^2.5.7",
-    "ionic2-super-tabs": "^4.1.4",
-    "node-sass": "^4.13.1",
-    "rxjs": "~6.5.1",
-    "tslib": "^1.9.0",
+    "cordova": "^9.0.0",
+    "cordova-plugin-stripe": "^1.5.3",
+    "core-js": "^2.6.11",
+    "ionic2-super-tabs": "^4.3.1",
+    "jshint": "^2.12.0",
+    "node-sass": "^4.14.1",
+    "pdfmake": "^0.1.68",
+    "rxjs": "^6.6.0",
+    "rxjs-compat": "^6.6.0",
+    "toolbox": "^0.1.1",
+    "tslib": "^1.13.0",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.801.2",
     "@angular-devkit/core": "~8.1.2",
     "@angular-devkit/schematics": "~8.1.2",
-    "@angular/cli": "~8.1.2",
+    "@angular/cli": "^8.3.28",
     "@angular/compiler": "~8.1.2",
     "@angular/compiler-cli": "~8.1.2",
     "@angular/language-service": "~8.1.2",
     "@capacitor/cli": "1.5.0",
-    "@ionic/angular-toolkit": "^2.1.1",
+    "@ionic/angular-toolkit": "^2.3.0",
     "@ionic/lab": "3.0.1",
+    "@types/chart.js": "^2.9.23",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^5.4.4",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "typescript": "^3.4.5"
   },
-  "description": "An Ionic project"
+  "description": "An Ionic project",
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-stripe": {}
+    }
+  }
 }
