@@ -1,4 +1,3 @@
-
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './gaurds/auth.guard';
@@ -6,6 +5,11 @@ import { AdminauthGuard } from './gaurds/adminauth.guard';
 
 
 const routes: Routes = [
+
+  {
+    path: 'menu',
+    loadChildren: () => import('./pages/vendor/menu/menu.module').then( m => m.MenuPageModule)
+  },
   {
     path: '',
     loadChildren: () => import('./index/index.module').then(m => m.IndexPageModule)
@@ -19,7 +23,7 @@ const routes: Routes = [
     loadChildren: () => import('./order/order.module').then( m => m.OrderPageModule)
   },
   {
-    path: 'cart-modal/:name',
+    path: 'cart-modal',
     loadChildren: () => import('./pages/cart-modal/cart-modal.module').then( m => m.CartModalPageModule)
   },
   {
@@ -50,7 +54,7 @@ const routes: Routes = [
     loadChildren: () => import('./pages/singupadmin/singupadmin.module').then( m => m.SingupadminPageModule)
   },
   {
-    path: 'signupcust',
+    path: 'singupcust',
 
     loadChildren: () => import('./pages/singupcust/singupcust.module').then( m => m.SingupcustPageModule)
   },
@@ -73,8 +77,16 @@ const routes: Routes = [
     loadChildren: () => import('./pages/vendor/add-menu/add-menu.module').then( m => m.AddMenuPageModule)
   },
   {
+    path: 'add-menu/:id/:name/:price/:description/:status',
+    loadChildren: () => import('./pages/vendor/add-menu/add-menu.module').then( m => m.AddMenuPageModule)
+  },
+  {
     path: 'overview',
     loadChildren: () => import('./pages/vendor/overview/overview.module').then( m => m.OverviewPageModule)
+  },
+  {
+    path: 'menu',
+    loadChildren: () => import('./pages/vendor/menu/menu.module').then( m => m.MenuPageModule)
   },
   {
 
@@ -96,18 +108,12 @@ const routes: Routes = [
   {
     path: 'login',
     loadChildren: () => import('./pages/login/login.module').then( m => m.LoginPageModule)
-
   },
 
   {
 
     path: 'history',
     loadChildren: () => import('./pages/history/history.module').then( m => m.HistoryPageModule)
-
-  },
-  {
-    path: 'signin',
-    loadChildren: () => import('./pages/signin/signin.module').then( m => m.SigninPageModule)
   },
   {
     path: 'custprof',
@@ -122,6 +128,55 @@ const routes: Routes = [
     path: 'reset-password',
     loadChildren: () => import('./pages/reset-password/reset-password.module').then( m => m.ResetPasswordPageModule)
   },
+  {
+    path: 'super-admin',
+    loadChildren: () => import('./pages/super-admin/super-admin.module').then( m => m.SuperAdminPageModule)
+  },
+  {
+    path: 'show-menu/:id/:name/:price/:description/:status',
+    loadChildren: () => import('./pages/vendor/show-menu/show-menu.module').then( m => m.ShowMenuPageModule)
+  },
+  {
+    path: 'vendor-menu1',
+    loadChildren: () => import('./pages/vendor/vendor-menu1/vendor-menu1.module').then( m => m.VendorMenu1PageModule)
+  },
+  {
+    path: 'vendor-menu3',
+    loadChildren: () => import('./pages/vendor/vendor-menu3/vendor-menu3.module').then( m => m.VendorMenu3PageModule)
+  },
+  {
+    path: 'notifications',
+    loadChildren: () => import('./pages/notifications/notifications.module').then( m => m.NotificationsPageModule)
+  },
+  {
+    path: 'payments',
+    loadChildren: () => import('./pages/payments/payments.module').then( m => m.PaymentsPageModule)
+  },
+  {
+    path: 'cart',
+    loadChildren: () => import('./pages/cart/cart.module').then( m => m.CartPageModule)
+  },
+   {
+     path: 'stripe',
+     loadChildren: () => import('./pages/stripe/stripe.module').then( m => m.StripePageModule)
+  },
+   {
+    path: 'stripe-web',
+    loadChildren: () => import('./pages/stripe-web/stripe-web.module').then( m => m.StripeWebPageModule)
+   },
+  {
+    path: 'edit',
+    loadChildren: () => import('./edit/edit.module').then( m => m.EditPageModule)
+  },
+  {
+    path: 'invoice/:id',
+    loadChildren: () => import('./invoice/invoice.module').then( m => m.InvoicePageModule)
+  },
+  {
+    path: 'landing',
+    loadChildren: () => import('./pages/landing/landing.module').then( m => m.LandingPageModule)
+  }
+
 
 ];
 @NgModule({
