@@ -8,16 +8,16 @@ import { AuthenticationService } from '../services/authentication.service';
   providedIn: 'root'
 })
 export class AuthGuard implements CanActivate {
- 
+
   constructor(private dataService: DataService, private router: Router, private authenticationService: AuthenticationService) { }
   canActivate(): boolean {
-   /* if (this.dataService.loggedIn()) { //from toolbox
+   if (this.dataService.loggedIn()) {
       return true;
     } else {
       this.router.navigate(['/signin']);
       return false;
-    }*/
-    return this.authenticationService.isAuthenticated();
+    }
+   return this.authenticationService.isAuthenticated();
 
    }
 }
