@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { NavController, NavParams, AlertController, ActionSheetController ,LoadingController} from '@ionic/angular';
+import { NavController, NavParams, AlertController, ActionSheetController , LoadingController} from '@ionic/angular';
 import { Router } from '@angular/router';
 import { DataService } from '../services/data.service';
 import { SearchService } from '../services/search.service';
@@ -11,15 +11,29 @@ import { SearchService } from '../services/search.service';
 })
 export class HomePage implements OnInit {
 
-  
+  contentLoaded = false;
    // tslint:disable-next-line: max-line-length
-   constructor(private route: Router,public actionSheetController: ActionSheetController,
-               private loadingCtrl: LoadingController,private dataService: DataService,
-               private searchService: SearchService) {}
+   constructor(private route: Router, public actionSheetController: ActionSheetController,
+               private loadingCtrl: LoadingController, private dataService: DataService,
+               private searchService: SearchService) {
+                 setTimeout(() => {
+                   this.contentLoaded = true;
+                 }, 3000);
+               }
 
-   searchTerm: any = "";
+   searchTerm: any = '';
 
-   // this a function of the actionsheet  
+  // testing the admin data
+  adminData: any = [];
+
+  data: any = [];
+
+  // search
+  // searchData: any = [];
+  rest: any = [];
+  jsonData: any = [];
+
+   // this a function of the actionsheet
 
    async presentActionSheet() {
      const actionSheet = await this.actionSheetController.create({
@@ -43,40 +57,42 @@ export class HomePage implements OnInit {
          text: 'Restaurants',
          icon: 'restaurant-outline',
          handler: () => {
-           console.log('Play clicked');
+           console.log('Restaurants clicked');
            this.route.navigateByUrl('/restaurants');
          }
        }, {
          text: 'My cart',
          icon: 'cart-outline',
          handler: () => {
-           console.log('Favorite clicked');
+           console.log('My cart clicked');
            this.route.navigateByUrl('/cart-modal');
          }
        },
         {
-         text: 'Register as a Vendor',
+         text: 'Signup',
          icon: 'person-add-outline',
          handler: () => {
-           console.log('Favorite clicked');
-           this.route.navigateByUrl('/signupadmin');
+           console.log('Signup clicked');
+           this.route.navigateByUrl('/signup');
          }
-       },
+        },
+        {
+          text: 'Order',
+          icon: 'time-outline',
+          handler: () => {
+            console.log('Share clicked');
+            this.route.navigateByUrl('/order2');
+          }
+        },
+      {
+       text: 'login',
+       icon: 'person',
+       handler: () => {
+         console.log('login clicked');
+         this.route.navigateByUrl('/login');
+       }
+   },
        {
-        text: 'Register as a customer',
-        icon: 'person-add-outline',
-        handler: () => {
-          console.log('Favorite clicked');
-          this.route.navigateByUrl('/signupcust');
-        }
-      },   {
-        text: 'Sign in',
-        icon: 'person',
-        handler: () => {
-          console.log('Favorite clicked');
-          this.route.navigateByUrl('/signin');
-        }
-      },{
         text: 'History',
         icon: 'time-outline',
         handler: () => {
@@ -90,7 +106,7 @@ export class HomePage implements OnInit {
           console.log('Share clicked');
           this.route.navigateByUrl('/profile');
         }
-      },{
+      }, {
         text: 'About Us',
         icon: 'information-circle-outline',
         handler: () => {
@@ -105,11 +121,20 @@ export class HomePage implements OnInit {
             console.log('admin clicked');
             this.route.navigateByUrl('/vendor-admin');
           }
-      },{
+      }
+      ,
+         {
+          text: 'goto superadmin',
+          icon: 'help-circle-outline',
+          handler: () => {
+            console.log('super admin clicked');
+            this.route.navigateByUrl('/super-admin');
+          }
+      }, {
          text: 'Cancel',
          icon: 'close',
          role: 'cancel',
-    
+
          handler: () => {
            console.log('Cancel clicked');
          }
@@ -117,61 +142,76 @@ export class HomePage implements OnInit {
      });
      await actionSheet.present();
    }
-  
-/*
-   //Loader controller function
-   presentLoading() {
-    const loader = this.loadingCtrl.create({
-      content: "Please wait...",
-      duration: 3000
-    });
-    loader.present();
-    \
-  }*/
 
-  // testing the admin data
-  adminData: any =[];
-
-  // search 
-  // searchData: any = []; 
-  jsonData: any = [];
-
-  public getRest(){ // ive used the adminData here to test the connection. some of the variables should be change..pls dont touch
-    return this.dataService.adminService().subscribe((data: any) => {this.adminData = data; console.log(this.adminData);});
+  public getRest() { // ive used the adminData here to test the connection. some of the variables should be change..pls dont touch
+    return this.dataService.adminService().subscribe((data: any) => {this.adminData = data; console.log(this.adminData); });
   }
 
   ngOnInit() {
     this.getRest();
   }
-  gotoRest(){
+  gotoRest() {
     this.route.navigateByUrl('/restaurants');
   }
 
-  // search bar
-  // --------------------------------------------------
-  ionViewDidLoad() {
-    this.setFilteredItems();
+  gotoMenu1() {
+
+    this.route.navigateByUrl('/menu');
+
+  }
+  // Refresh fuction
+  doRefresh(event) {
+
+    setTimeout(() => {
+      this.ionViewWillEnter();
+      event.target.complete();
+    }, 550);
+
   }
 
-  setFilteredItems() {
- 
-        this.jsonData = this.dataService.filterItems(this.searchTerm);
-       // this.searchData = this.searchService.filterItems(this.searchTerm);
- 
-    }
-// ------------------------------------------------------
+  gotoVendorMenu2() {
+    this.route.navigateByUrl('/order2');
+  }
+
+  gotoVendorMenu3() {
+    this.route.navigateByUrl('/order2');
+  }
+  // search bar
+  // --------------------------------------------------
+//   ionViewDidLoad() {
+//    this.setFilteredItems();
+// }
 
 
-
-    search(){
-      this.dataService.jsonData.subscribe((dat: any) => {this.jsonData = dat; console.log(this.jsonData);});
-     // return this.searchService.searchD().subscribe((dat: any) => {this.searchData = dat;console.log(this.searchData);});
-    }
-
-    slidesDidLoad(slides) {
-      slides.startAutoplay();
-    }
-  
-
-
+  ionViewWillEnter() {
+    setTimeout(() => {
+      this.data = {
+        heading: 'mama`s kitchen',
+        para1: 'real african food',
+        para2: 'food for thought.'
+      };
+    }, 5000);
+  }
 }
+
+// function for the restaurant click to menu
+
+
+//  setFilteredItems() {
+
+//          this.jsonData = this.dataService.filterItems(this.searchTerm);
+//         /this.searchData = this.searchService.filterItems(this.searchTerm);
+// }
+// // ------------------------------------------------------
+
+
+
+//     search() {
+//       this.dataService.jsonData.subscribe((dat: any) => {this.jsonData = dat; console.log(this.jsonData); });
+//      // return this.searchService.searchD().subscribe((dat: any) => {this.searchData = dat;console.log(this.searchData);});
+//     }
+
+//     slidesDidLoad(slides) {
+//       slides.startAutoplay();
+//     }
+
