@@ -10,7 +10,7 @@ const routes: Routes = [
         {
             path: 'feed',
             loadChildren: () => import('../pages/feed/feed.module').then(
-                m => m.FeedPageModule
+                m                       => m.FeedPageModule
             )
         },
         {
