@@ -6,7 +6,6 @@ import { CartModalPage } from '../pages/cart-modal/cart-modal.page';
 import { ExtrasPage } from '../pages/extras/extras.page';
 import { Location } from '@angular/common';
 
-
 @Component({
   selector: 'app-order',
   templateUrl: './order.page.html',
@@ -15,61 +14,72 @@ import { Location } from '@angular/common';
 export class OrderPage implements OnInit {
   cart = [];
   products = [];
+  ext = [];
   cartItemCount: BehaviorSubject<number>;
+  menuList: Array<any> = [];
 
-  @ViewChild('cart',{static:false,read: ElementRef})fab: ElementRef;
+  @ViewChild('cart', {static: false, read: ElementRef})fab: ElementRef;
 
-  constructor(private cartService: CartService, private modalCtrl: ModalController,private location: Location) { }
+  constructor(private cartService: CartService, private modalCtrl: ModalController, private location: Location) { }
 
+  // ngOnInit() {
+  //   this.products = this.cartService.getProducts();
+  //   this.cart = this.cartService.getCart();
+  //   this.cartItemCount = this.cartService.getCartItemCount();
+  // }
   ngOnInit() {
-    this.products = this.cartService.getProducts();
-    this.cart = this.cartService.getCart();
-    this.cartItemCount = this.cartService.getCartItemCount();
+
+    // this.cartService.getItems().subscribe((data: any) => {
+    //   this.menuList = data.menu;
+    //   console.log(this.menuList);
+    // });
+    // this.ext = this.cartService.getExt();
+    // this.cartItemCount = this.cartService.getCartItemCount();
   }
 
-  addToCart(product){
+
+  addToCart(product) {
     this.animateCSS('tada');
     this.cartService.addProduct(product);
   }
-  exProduct(product){
+  exProduct(product) {
     this.cartService.extraProd(product);
   }
-  async openCart(){
-    this.animateCSS('bounceOutLeft',true);
-    let modal = await this.modalCtrl.create({
+  async openCart() {
+    this.animateCSS('bounceOutLeft', true);
+    const modal = await this.modalCtrl.create({
       component: CartModalPage,
       cssClass: 'cart-modal'
     });
     modal.onWillDismiss().then(() => {
-      this.fab.nativeElement.classList.remove('animated','bounceOutLeft')
+      this.fab.nativeElement.classList.remove('animated', 'bounceOutLeft');
       this.animateCSS('bounceLeft');
     });
     modal.present();
   }
-  async openExtras(){
-    let modal = await this.modalCtrl.create({
+  async openExtras() {
+    const modal = await this.modalCtrl.create({
       component: ExtrasPage,
       cssClass: 'extras'
     });
     modal.present();
-    
   }
 
-  animateCSS(animationName, keepAnimated = false){
+  animateCSS(animationName, keepAnimated = false) {
     const node = this.fab.nativeElement;
-    node.classList.add('animated',animationName)
+    node.classList.add('animated', animationName);
 
-    function handleAnimationEnd(){
-      if(!keepAnimated){
-        node.classList.remove('animated',animationName);
+    function handleAnimationEnd() {
+      if (!keepAnimated) {
+        node.classList.remove('animated', animationName);
       }
-      node.removeEventListener('animationend',handleAnimationEnd)
+      node.removeEventListener('animationend', handleAnimationEnd);
     }
-    node.addEventListener('animationend',handleAnimationEnd)
+    node.addEventListener('animationend', handleAnimationEnd);
   }
 
-  //added the back button manipulation-Tumie
-  backButton(){
+  // added the back button manipulation-Tumie
+  backButton() {
     this.location.back();
    }
 
