@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router} from '@angular/router';
 
 @Component({
   selector: 'app-login',
@@ -7,9 +8,17 @@ import { Component, OnInit } from '@angular/core';
 })
 export class LoginPage implements OnInit {
 
-  constructor() { }
+  // tslint:disable-next-line: variable-name
+  user_name: string;
+  constructor(private router: Router ) { }
 
   ngOnInit() {
+  }
+
+  gotoSignup() {
+
+      this.router.navigateByUrl('signup');
+
   }
 
 }
