@@ -2,11 +2,12 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { CartService } from 'src/app/services/cart.service';
+import { Product } from '../../services/cart.service';
 import { ModalController } from '@ionic/angular';
 import { CartModalPage } from '../cart-modal/cart-modal.page';
-
+import { Router } from '@angular/router';
 import { ExtrasPage } from '../extras/extras.page';
-
+import { PostProvider } from '../../../providers/post-provider';
 import { Location } from '@angular/common';
 
 
@@ -17,65 +18,58 @@ import { Location } from '@angular/common';
 })
 export class Order2Page implements OnInit {
 
+  menuItems: any[];
+  limit = 10;
+  start = 0;
   cart = [];
   product = [];
+  products = [];
   cartItemCount: BehaviorSubject<number>;
 
-  @ViewChild('cart',{static:false,read: ElementRef})fab: ElementRef;
+  @ViewChild('cart', {static: false, read: ElementRef})fab: ElementRef;
 
-  constructor(private cartService: CartService, private modalCtrl: ModalController, private location: Location) { }
+  // tslint:disable-next-line: max-line-length
+  constructor(private cartService: CartService, private modalCtrl: ModalController, private router: Router, private postPvdr: PostProvider ) {}
 
   ngOnInit() {
-    this.product = this.cartService.getProduct();
+    this.products = this.cartService.getProducts();
     this.cart = this.cartService.getCart();
     this.cartItemCount = this.cartService.getCartItemCount();
+
   }
 
-  addToCart(product){
-    this.animateCSS('tada');
+  addToCart(product) {
     this.cartService.addProduct(product);
+    this.animateCSS('tada');
   }
-  exProduct(product){
-    this.cartService.extraProd(product);
-  }
-  async openCart(){
-    this.animateCSS('bounceOutLeft',true);
-    let modal = await this.modalCtrl.create({
+
+  async openCart() {
+    this.animateCSS('bounceOutLeft', true);
+
+    const modal = await this.modalCtrl.create({
       component: CartModalPage,
       cssClass: 'cart-modal'
     });
     modal.onWillDismiss().then(() => {
-      this.fab.nativeElement.classList.remove('animated','bounceOutLeft')
-      this.animateCSS('bounceLeft');
+      this.fab.nativeElement.classList.remove('animated', 'bounceOutLeft');
+      this.animateCSS('bounceInLeft');
     });
     modal.present();
   }
 
-  async openExtras(){
-    let modal = await this.modalCtrl.create({
-      component: ExtrasPage,
-      cssClass: 'extras'
-    });
-    modal.present();
-    
-  }
-
-  animateCSS(animationName, keepAnimated = false){
+  animateCSS(animationName, keepAnimated = false) {
     const node = this.fab.nativeElement;
-    node.classList.add('animated',animationName)
+    node.classList.add('animated', animationName);
 
-    function handleAnimationEnd(){
-      if(!keepAnimated){
-        node.classList.remove('animated',animationName);
+    function handleAnimationEnd() {
+      if (!keepAnimated) {
+        node.classList.remove('animated', animationName);
       }
-      node.removeEventListener('animationend',handleAnimationEnd)
+      node.removeEventListener('animationend', handleAnimationEnd);
     }
-    node.addEventListener('animationend',handleAnimationEnd)
+    node.addEventListener('animationend', handleAnimationEnd);
   }
-
-//back button
-backButton(){
-  this.location.back();
- }
 
 }
+
+
