@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import 'rxjs/add/operator/map';
 
 import { RestaurantService } from '../services/restaurant.service';
 
 import { Location } from '@angular/common';
-//import { RestaurantService } from '../services/restaurant.service';
-//import { ActionsheetComponent } from '../components/actionsheet/actionsheet.component';
+import { DataService } from '../services/data.service';
+// import { RestaurantService } from '../services/restaurant.service';
+// import { ActionsheetComponent } from '../components/actionsheet/actionsheet.component';
 
 
 
@@ -17,29 +19,33 @@ import { Location } from '@angular/common';
 export class RestaurantsPage implements OnInit {
 
 
+  constructor(private route: Router, private location: Location, private dataService: DataService) { }
+rest: any = [];
 
-  constructor(private route: Router, private location: Location) { }
-
-   
   ngOnInit() {
-    
+   this.getRest();
   }
  /* list()
   {
     return this.asheet.presentActionSheet()
   }*/
- order(){
+ order() {
    this.route.navigateByUrl('/order2');
  }
- secOrder(){
+ secOrder() {
    this.route.navigateByUrl('/order');
  }
- thirdOrder(){
+ thirdOrder() {
    this.route.navigateByUrl('/order3');
  }
 
- backButton(){
+ backButto() {
   this.location.back();
+ }
+ getRest() {
+
+  return this.dataService.grest()
+  .subscribe(data => {this.rest = data ; console.log(data); });
  }
 
 }
